Release Q2.2.1
Server - Implemented a High Sampling Rate option for the CHS42X4, ICS421 and ICS425 Modules. In this Operation mode the Module will disable channels 2 and 5 allowing the remaining 4 channels to allocate at an effective Sampling Rate of MSR divide by 1.
Server - Improved buffer level calculations. This adds a bit more stability if there is jitter on the client streaming application.